+++ pkg/noun/jets/b/lent.c
@@ -6,7 +6,7 @@
 
 #include "noun.h"
 
-STATIC_ASSERT( (UINT32_MAX > u3a_cells),
+STATIC_ASSERT( (UINT64_MAX > u3a_cells),
                "length precision" );
 
 u3_noun
